requirements.txt
flask
requests
python-decouple
supabase
pyjwt